Align maintenance window comment with Service Description
The Service Description for IBM Cloud Platform (http://www-03.ibm.com/software/sla/sladb.nsf/pdf/6605-12/$file/i126-6605-12_11-2017_en_US.pdf) makes no reference to a 24 notification window for planned outages. Contractually, IBM makes no guarantee on the lead time for such notifications but in reality non-emergency maintenance is announced well in advance. This reference to 24 hours (which is low) is a little misleading and this bullet should align with the service description.